Update patch set 1
Patch Set 1:

I'm not sure if we're talking about the same thing, but I don't think that this "speed bump" is working. Scrolling between homescreens stutters like crazy (somewhere around or below 15 fps for sure) on a Galaxy Nexus.

It's obviously meant to be a slow wallpaper so 15 fps didn't matter when it was static, but making it scrollable seems to drag down the fps of the entire launcher.

Patch-set: 1