Update patch set 2
Patch Set 2:

The use case is for custom OOBEs on shipping devices. I don't think we should be doing this across all devices of course, but changing the colors of something like this is vulnerable to a six month debate about the absolute correct hue and saturation :) People using CM to build devices can have a little flexibility with this patch.

Patch-set: 2